# Add object state cache for moveos