--- a/Cargo.toml
+++ b/Cargo.toml
@@ -267,6 +267,7 @@ tower_governor = { version = "0.4.3", features = ["tracing"]}
 pin-project = "1.1.7"
 mirai-annotations = "1.12.0"
 lru = "0.11.0"
+quick_cache = "0.6.9"
 bs58 = "0.5.1"
 dirs-next = "2.0.0"
 anstream = { version = "0.3" }