--- a/moveos/moveos-store/Cargo.toml
+++ b/moveos/moveos-store/Cargo.toml
@@ -32,3 +32,4 @@ raw-store = { workspace = true }
 moveos-config = { workspace = true }
 accumulator = { workspace = true }
 metrics = { workspace = true }
+quick_cache = { workspace = true }