## Summary

[deprecated] Add object state cache for moveos. By querying 100 identical object_ids with rooch_queryObjectStates simultaneously, the client response time was reduced from 1.72s to 0.38s.

### version 2

1. only keep state cache and move cache to StateDBStore.

performance:
1. By querying 100 identical object_ids with rooch_queryObjectStates simultaneously, the client response time was reduced from 1.17s to 0.51s
2.  rooch benchmark:
![image](https://github.com/user-attachments/assets/7691fa2f-f317-4755-a1ed-be45560509fb)